+++ example/rest-server.php
@@ -25,7 +25,7 @@ use RestDeamon\Example\Endpoint\Index\GetHandler;
 use RestDeamon\Example\Endpoint\Reusable\HelloHandler;
 
 $httpDriverClass = \FreeElephants\RestDaemon\HttpDriver\Aerys\AerysDriver::class;
-$server = new RestServer('127.0.0.1', 8080, '0.0.0.0', ['*']);
+$server = new RestServer('127.0.0.1', 8080, '0.0.0.0', ['*'], $httpDriverClass);
 
 $requestCounter = function (
     ServerRequestInterface $request,