+++ src/FreeElephants/RestDaemon/HttpAdapter/Aerys2Zend/ServerRequest.php
@@ -3,6 +3,7 @@
 namespace FreeElephants\RestDaemon\HttpAdapter\Aerys2Zend;
 
 use Aerys\Request;
+use Psr\Http\Message\UriInterface;
 use Zend\Diactoros\ServerRequest as ZendServerRequest;
 use Zend\Diactoros\Uri;
 
@@ -16,13 +17,33 @@ class ServerRequest extends ZendServerRequest
     {
         $serverParams = $_SERVER;
         $uploadedFiles = [];
-        $uri = new Uri($request->getUri());
+        $uri = $this->mapUri($request);
         $method = $request->getMethod();
         $headers = $request->getAllHeaders();
         $cookies = [];
         $queryParams = $request->getAllParams();
         $parsedBody = [];
         $protocol = $request->getProtocolVersion();
-        parent::__construct($serverParams, $uploadedFiles, $uri, $method, 'php://memory', $headers, $cookies, $queryParams, $parsedBody, $protocol);
+        parent::__construct(
+            $serverParams,
+            $uploadedFiles,
+            $uri,
+            $method,
+            'php://memory',
+            $headers,
+            $cookies,
+            $queryParams,
+            $parsedBody,
+            $protocol);
+    }
+
+    private function mapUri(Request $request): UriInterface
+    {
+        $uri = new Uri($request->getUri());
+        $connectionInfo = $request->getConnectionInfo();
+        $scheme = $connectionInfo['is_encrypted'] ? 'https' : 'http';
+        $host = $connectionInfo['server_addr'];
+        $port = $connectionInfo['server_port'];
+        return $uri->withScheme($scheme)->withHost($host)->withPort($port);
     }
 }