+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - EndpointMethodHandlerInterface::handle() turned to final and should be changed to private.  
 
 ### Fixed
-- Request scheme and host now available in handlers and middleware (see Ratchet fix https://github.com/ratchetphp/Ratchet/pull/471) 
+- Request scheme and host now available in handlers and middleware (see Ratchet fix https://github.com/ratchetphp/Ratchet/pull/471). Aerys driver fixed too. 
 
 ### Internal
 - Extend CallableEndpointMethodHandlerWrapper from AbstractEndpointMethodHandler.